requirements.txt
pyserial>=3.5
SpeechRecognition>=3.10.0
google-generativeai>=0.3.0
PyAudio>=0.2.11